refactor: QuantTensor
