+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/bigquant/SpatialDilatedConvolution.scala
@@ -16,7 +16,7 @@
 
 package com.intel.analytics.bigdl.nn.bigquant
 
-import com.intel.analytics.bigdl.tensor.QuantTensor
+import com.intel.analytics.bigdl.tensor.{QuantTensor, Tensor}
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.serializer.{DataConverter, ModuleData}
 import scala.reflect.ClassTag
@@ -75,45 +75,18 @@ object SpatialDilatedConvolution extends QuantSerializer {
 
   override def serializeWeight[T: ClassTag](module: ModuleData[T],
     modelBuilder: BigDLModule.Builder)(implicit ev: TensorNumeric[T]): Unit = {
-    val conv = module.module.asInstanceOf[SpatialConvolution[T]]
-    val offsets = new Array[Int](conv.weight.length)
-    val allWeights = new Array[Byte](conv.nOutputPlane * conv.nInputPlane *
-            conv.kernelH * conv.kernelW / conv.nGroup)
-
-    var currentOffset = 0
-    for (i <- conv.weight.indices) {
-      offsets(i) = conv.weight(i).size().product
-      System.arraycopy(conv.weight(i).getStorage, 0, allWeights, currentOffset, offsets(i))
-      currentOffset += offsets(i)
-    }
-
-    val offsetBuilder = AttrValue.newBuilder
-    DataConverter.setAttributeValue(offsetBuilder, offsets, universe.typeOf[Array[Int]])
-    modelBuilder.putAttr("offsets", offsetBuilder.build)
-
+    val conv = module.module.asInstanceOf[SpatialDilatedConvolution[T]]
     val weightBuilder = AttrValue.newBuilder
-    DataConverter.setAttributeValue(weightBuilder, allWeights, universe.typeOf[Array[Byte]])
+    DataConverter.setAttributeValue(weightBuilder, conv.weight)
     modelBuilder.putAttr("weights", weightBuilder.build)
   }
 
   override def loadWeight[T: ClassTag](model: BigDLModule,
     module: ModuleData[T])(implicit ev: TensorNumeric[T]): Unit = {
-    val conv = module.module.asInstanceOf[SpatialConvolution[T]]
+    val conv = module.module.asInstanceOf[SpatialDilatedConvolution[T]]
     val attrMap = model.getAttrMap
-    val offsets = DataConverter.getAttributeValue(attrMap.get("offsets")).asInstanceOf[Array[Int]]
-    val byteArray = DataConverter.getAttributeValue(attrMap.get("weights"))
-            .asInstanceOf[Array[Byte]]
-
-    var currentOffset = 0
-    conv.weight = new Array[QuantTensor[T]](offsets.length)
-    for (i <- conv.weight.indices) {
-      conv.weight(i) = new QuantTensor[T](conv.nOutputPlane / conv.nGroup,
-        conv.nInputPlane / conv.nGroup, conv.kernelH, conv.kernelW)
-      val storage = new Array[Byte](conv.weight(i).size().product)
-      System.arraycopy(byteArray, currentOffset, storage, 0, offsets(i))
-      currentOffset += offsets(i)
-      conv.weight(i).setStorage(storage)
-    }
+    conv.weight = DataConverter.getAttributeValue(attrMap.get("weights"))
+      .asInstanceOf[Array[Tensor[T]]]
   }
 }
